sun and moon are moving problem with the sizes (glPerspective) not resolved

+++ src/main/java/com/momega/spacesimulator/controller/CameraController.java
@@ -33,11 +33,11 @@ public class CameraController extends AbstractController {
                 break;
 
             case KeyEvent.VK_Q:
-                camera.twist(+0.5f);
+                camera.twist(+0.5);
                 break;
 
             case KeyEvent.VK_E:
-                camera.twist(-0.5f);
+                camera.twist(-0.5);
                 break;
 
             case KeyEvent.VK_S:
@@ -45,19 +45,19 @@ public class CameraController extends AbstractController {
                 break;
 
             case KeyEvent.VK_O:
-                camera.lookLeft(+0.5f);
+                camera.lookLeft(+0.5);
                 break;
 
             case KeyEvent.VK_P:
-                camera.lookLeft(-0.5f);
+                camera.lookLeft(-0.5);
                 break;
 
             case KeyEvent.VK_H:
-                camera.lookUp(+0.5f);
+                camera.lookUp(+0.5);
                 break;
 
             case KeyEvent.VK_N:
-                camera.lookUp(-0.5f);
+                camera.lookUp(-0.5);
                 break;
         }
     }
@@ -76,8 +76,8 @@ public class CameraController extends AbstractController {
 
         Point delta = new Point(e.getX() - mouseLast.x, e.getY() - mouseLast.y);
 
-        final float MOUSE_SPEED_MODIFIER = 0.25f;
-        float horizAngle = delta.x*MOUSE_SPEED_MODIFIER, vertAngle = delta.y*MOUSE_SPEED_MODIFIER;
+        final double MOUSE_SPEED_MODIFIER = 0.25f;
+        double horizAngle = delta.x*MOUSE_SPEED_MODIFIER, vertAngle = delta.y*MOUSE_SPEED_MODIFIER;
 
         // Turn horizontally by rotating about the standard up vector (0,0,1).
         camera.lookLeft(-horizAngle);
